+++ doc/examples/render_flip/lib/main.dart
@@ -1,6 +1,5 @@
 import 'package:flame/sprite_animation.dart';
 import 'package:flame/components/sprite_animation_component.dart';
-import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/extensions/vector2.dart';
 import 'package:flutter/material.dart' hide Image;
@@ -8,8 +7,7 @@ import 'dart:ui';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
-  final Vector2 size = await Flame.util.initialDimensions();
-  final game = MyGame(size);
+  final game = MyGame();
   runApp(
     GameWidget(
       game: game,
@@ -55,9 +53,5 @@ class MyGame extends BaseGame {
         stepTime: 0.15,
       ),
     );
-  }
-
-  MyGame(Vector2 screenSize) {
-    size.setFrom(screenSize);
   }
 }