+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - Fix update+render of component children
  - Update documentation for SVG component
  - Adding Component#onLoad
+ - Moving size to Game instead of BaseGame
 
 ## 1.0.0-rc2
  - Improve IsometricTileMap and Spritesheet classes